pyproject.toml
[tool.poetry]
name = "chat-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
keywords = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"  # Supports Python 3.10 to <4.0
streamlit = "*"
python-dotenv = "*"
langchain-openai = "*"
langchain-groq = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

